requirements.txt
numpy~=1.22.3
Pillow~=9.1.1
scipy~=1.8.1
matplotlib~=3.5.2